requirements.txt
python-dotenv~=1.0.0
pydantic~=1.10.12
fastapi~=0.100.0
torch~=2.0.1
transformers~=4.30.2
tiktoken~=0.4.0
pypdf2~=3.0.1
langchain~=0.0.276
pymongo~=4.4.1
faiss-cpu==1.7.4
python-multipart==0.0.6